fix : seq의 타입을 int에서 long으로 변경

+++ src/main/java/someone/alcoholic/domain/Board/Board.java
@@ -22,7 +22,7 @@ import java.util.List;
 public class Board {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int seq;
+    private Long seq;
 
     @Column(nullable = false)
     private String title;
